Propose a template architecture, v2
There were some problem with the previous proposal; in particular, Ansible
does not allow us to provide all module arguments from a dictionary
variable, and this made it difficult to effectively make use of the
`templateParameters` dictionary in the cluster order.

With this commit, template modules now have a `template_parameters` option,
which takes a dictionary value. We can pass the value of
`templateParameters` from the ClusterOrder directly to this variable.

By locating template parameters in an option like this, we also have the
ability to specify role arguments that aren't provided via template
parameters, such as the node requests.

There are two example templates provided in this commit to show how this
might look.