Run post-delete job to report status
This replaces the "delete-hosted-cluster" job with a workflow
("delete-hosted-cluster-workflow") that runs a status reporting job after
the main job has finished. The status jobs are responsible for updating the
clusterorder with the success or failure of the delete operation.

This can be useful both to expose the status to the operator (from where we
could in theory publish it to the fulfillment service) and to possibly
re-trigger the delete webhook.

Currently the status is reported as an annotation on the clusterorder,
rather than as a status condition, because the kubernetes ansible modules
do not permit updates to the status subresource of kubernetes objects. See
[1] for details.

[1]: https://github.com/ansible-collections/kubernetes.core/issues/607

Part of: https://github.com/innabox/issues/issues/239